Try to fix hover flicker
There is a weird bug where hovering on just the right pixel of the collapse link
will cause the mouse cursor to flicker. It seems like changing the opacity
somehow changes the hover box of the link too, so that it shrinks/expands, and
cancels the hover, which then disables the opacity, which causes a hover again,
and so on. Adding a bit of padding seems to fix this.

With this padding, we can also remove the space before the mid-dot.