Fix collapse icon toggling
We can't just toggle the class. Suppose comment B replies to comment A.
We collapse A, which collapses B. We then expand B. With the old code,
now expanding A would toggle the collapse icon of B, resulting in an
expanded comment with a plus collapse icon. So, we must force the toggle.