Fix quote link hover flicker
Previously, it seems the popover comment overlapped the bottom of the
quote link's box. So, when the mouse was hovered at the very bottom of
the quote link, the popover comment would appear over the quote link,
and cause a mouseout. This then caused the popover to disappear, causing
a mouseover, and so on. A 3px gap seems to be enough to fix this.