Remove example for "when code must be copied"
It's not strictly true that it only happens when a big feature must be
added. In the case of beginShowQuote, we need to use cloneComment
instead of cloneNode. That call occurs in the middle of the function, so
it's much easier to copy the method and change one line than it is to
somehow wrap that. We also want to change an offset from 20px to 23px.
Again, it's easier to just do that directly than to edit it after the
normal code runs. Both methods are equally brittle anyway.