Clarify infinite expanding
We only check for quote links to the parent, and not to all ancestors.
That's probably not necessary, and more complex. So, here's a scenario
in which you could have infinite expansion:

A <- B <- C
^         |
+---------+

B replies to A, and C replies to B and A. Expand B's callback in A.
Then, expand C's callback in B. Now, expand A's quote link in C. This
achieves a cycle without ever expanding a parent.

Again, I imagine this exact reply structure is rare, and preventing it
would require searching up the tree for every expansion, so I don't
think it's worth checking for.