Don't reuse quote container comment
It can cause the wrong comment to be inserted. Scenario: expand a link.
Hover over any other link, replacing the comment in the quote container.
Hover over the original link. It's disabled, so it won't replace the
other comment in the container. Now, click the first link twice. This
removes the comment, and, without replacing the comment in the
container (no mousemove occurred), adds the comment from the container.
Thus, the wrong comment is added.