Call setupQuotes on init
On mobile, we need to call our setupQuotes to fix the links on page 1.
But, because we init late to avoid bugs, the original setupQuotes has
probably already been called. So we call just to be sure. This is a bit
inefficient (it calls the original setupQuotes again), but it's fine.