Full width hover comments
This forces your eye to jump to keep reading, but it's better than
making the comment very small when the nesting gets deep. The best
solution would be to align the .data of the hover comment with that of
the parent, but that needs precise CSS fiddling, which is too fragile.