Rename Comments Plus to Comment Tweaks
My original goal was to add a lot more features. The big idea was to
parse comments and provide a threaded view. But, I've recently spent
less time on Fimfiction, and don't have an interest in attempting that
anymore. This userscript is just a collection of things that I want, and
not a comprehensive upgrade. So, the name "tweaks" makes more sense.