Fix comment collapsing
Previously, comment collapsing worked, but it selected all quote
links--including those of expanded comments. I didn't notice any
problems with it for a while, but eventually it seemed like there were
some infinite loop bugs or something.

The new strategy for ASC is similar. The main difference is that we
select .comment_callbacks by ID (removing the ID on clones) and use a
child combinator. This just selects the top-level backlinks.

The strategy for DESC is more complicated. I couldn't find a good way to
select just the top-level quote links in .data. The only thing
separating those quote links from the quote links of expanded comments
is that one is a descendant of a .inline-quote. Since you can't :not()
descendant selectors, that can't work. Another approach might be to
extract the link data and store it like the indexes, but I'm afraid this
will lead to more complexity.

The final strategy looks at backlinks which point to the current
comment. That is, the backlinks which correspond to the links in .data.
We then find the parent comments of these backlinks, which are the
comments we need to collapse. I don't think it's that efficient, but it
does only rely on DOM lookups, which is simpler.