Fix/replace getCommentIndexRange
Deleted comments at the beginning or end of the page break it. We could
find the indexes of the first and last non-deleted comments, but that
means storeComments needs to start counting from the first non-deleted
comment. It's easier to just have storeComments and rewriteQuoteLinks do
DOM requests instead of checking against an index range. More DOM
operations, but less complexity (and less bugs).