Avoid document.getElementById
The bug was in rewriteQuoteLinks--hidden comments do have an ID, which
means that a comment which is first viewed as a foreign comment, and
later viewed normally, and then later expanded will not be rewritten.

For good measure, we avoid using document.getElementById when possible,
as it will probably cause problems with hidden comments.