Add back delay of 85ms
I thought I was clever by removing the delay. Turns out, it actually had
a purpose! Without the delay and canceling, if you quickly mouseover and
mouseout a link to a foreign comment, the request won't succeed until
after your mouse has left the link. Thus, there will be a floating
quote, as the mouseout event was triggered before the quote even
appeared. This adds back the ability to cancel showing the quote if the
user's mouse has already left the link. It also adds in a delay of 85ms,
which is fast enough that I don't notice it, and still provides some
buffer to guard against errant mouseovers.