Put hover comment under quote link
If a quote link's icon and text happen to break across two lines,
putting your cursor over the text wrapped to the second line will cause
a flicker: the hovered comment appears, causing a mouseout, which makes
it disappear, causing a mouseover, etc. Putting the hovered comment
under the bottom of the link makes it a bit far away in this case, but
it's better than a flicker I think.