Assign ctCSS before calling init()
It turns out I didn't understand variable hoisting as well as I thought
I did. Only the _declaration_ is hoisted--the assignment is left in
place. So, previously if init() was called in the if statement, ctCSS
would be undefined, which means that comment collapsing would break
(since the ct--collapsed-comment class had no styles attached to it).